Project layers setup
