+++ src/SchedulerManager/Travely.SchedulerManager.API/Services/GreeterService.cs
@@ -1,13 +1,12 @@
-using System.Threading.Tasks;
 using Grpc.Core;
 using Microsoft.Extensions.Logging;
+using System.Threading.Tasks;
 
-namespace Travely.SchedulerManager.Service.Services
+namespace Travely.SchedulerManager.API
 {
     public class GreeterService : Greeter.GreeterBase
     {
         private readonly ILogger<GreeterService> _logger;
-
         public GreeterService(ILogger<GreeterService> logger)
         {
             _logger = logger;
