+++ src/SchedulerManager/Travely.SchedulerManager.Abstraction/Travely.SchedulerManager.Abstraction.csproj
