--- src/SchedulerManager/Travely.SchedulerManager.Service/Properties/launchSettings.json
@@ -1,13 +1,0 @@
-﻿{
-  "profiles": {
-    "Travely.SchedulerManager.Service": {
-      "commandName": "Project",
-      "dotnetRunMessages": "true",
-      "launchBrowser": false,
-      "applicationUrl": "http://localhost:5000;https://localhost:5001",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
